requirements.txt
pycryptodome
qrcode
fpdf
opencv-python
Pillow